+++ src/mpsc.rs
@@ -106,7 +106,7 @@ impl<T> FlavorBounded for Array<T> {
     }
 }
 
-impl<T: 'static> Flavor for Array<T> {
+impl<T: Send + 'static> Flavor for Array<T> {
     type Send = RegistryMulti;
     type Recv = RegistrySingle;
 }
@@ -160,7 +160,7 @@ where
 #[inline]
 fn unbounded_new<T, R>() -> (MTx<List<T>>, R)
 where
-    T: 'static,
+    T: Send + 'static,
     R: ReceiverType<Flavor = List<T>> + NotCloneable,
 {
     build::<List<T>, MTx<List<T>>, R>(List::<T>::from_inner(crate::flavor::List::<T>::new()))
@@ -169,7 +169,7 @@ where
 #[inline]
 pub fn unbounded_blocking<T>() -> (MTx<List<T>>, Rx<List<T>>)
 where
-    T: 'static,
+    T: Send + 'static,
 {
     unbounded_new()
 }
@@ -177,14 +177,14 @@ where
 #[inline]
 pub fn unbounded_async<T>() -> (MTx<List<T>>, AsyncRx<List<T>>)
 where
-    T: 'static,
+    T: Send + 'static,
 {
     unbounded_new()
 }
 
 fn bounded_new<T, S, R>(size: usize) -> (S, R)
 where
-    T: 'static,
+    T: Send + 'static,
     S: SenderType<Flavor = Array<T>> + Clone,
     R: ReceiverType<Flavor = Array<T>> + NotCloneable,
 {
@@ -197,7 +197,7 @@ where
 #[inline]
 pub fn bounded_blocking<T>(size: usize) -> (MTx<Array<T>>, Rx<Array<T>>)
 where
-    T: 'static,
+    T: Send + 'static,
 {
     bounded_new(size)
 }
@@ -208,7 +208,7 @@ where
 #[inline]
 pub fn bounded_async<T>(size: usize) -> (MAsyncTx<Array<T>>, AsyncRx<Array<T>>)
 where
-    T: 'static,
+    T: Send + 'static,
 {
     bounded_new(size)
 }
@@ -219,7 +219,7 @@ where
 #[inline]
 pub fn bounded_blocking_async<T>(size: usize) -> (MTx<Array<T>>, AsyncRx<Array<T>>)
 where
-    T: 'static,
+    T: Send + 'static,
 {
     bounded_new(size)
 }
@@ -230,7 +230,7 @@ where
 #[inline]
 pub fn bounded_async_blocking<T>(size: usize) -> (MAsyncTx<Array<T>>, Rx<Array<T>>)
 where
-    T: 'static,
+    T: Send + 'static,
 {
     bounded_new(size)
 }